Do not evaluate the whole config in “overlays-compat”
Previous implementation makes every call of “nix-shell” or other Nix
stuff to evaluate my whole Nix configuration, every single time. Despite
that is makes it slower it also causes other problems. For instance when
I’m working on something in my NixOS config, and it’s broken (just WIP),
and when I just call some Nix stuff (also for the need of making it
done) it fails. Because there are errors in my configuration that I
didn’t even try to rebuild yet. This is annoying of course.

By simplifying the implementation I’m fixing this problem. The config
isn’t going to be re-evaluated everytime anymore. This may drop some
features, like, as I understand, it would ignore
“~/.config/nixpkgs/config.nix” which I don’t use anyway.
So I just hope current solution would be sufficient for me.