Pointer script: Support devices with the same name
Previously the generic pointer script was only taking first match for
the given device name. Sometimes you get few device xinput items with
the same name for the same device. In this case the script doesn’t work
when the device you need to change property value for is not the first
one in the list xinput gives you.

This change brings a huge refactor for the script which fixes the
original problem and adds some helpers for ensuring the given property
actually exists (the existence of the property is what matters, not only
that the device is found by its name). Technically this script can work
now not only with pointers but also with keyboards if it’s ever needed.