XMonad config: Use “place-cursor-at” only for jumping between screens
My “cursor-to-display” script is mismatching the native XMonad order of
the screens. Depending on how complex the layout is when pressing m+uiop
one screen gets focused (a window on it) but the mouse cursor jumps to a
different screen. “place-cursor-at” gets information about the screens
order from Xinerama which is matching XMonad’s.