Simplify “build-audio-profile.sh” script
Remove another call of `nixos-rebuild` for restoring the default profile
as default one (last one built) for Grub to run. Also remove `sudo`
wrapper (it’s not necessary for `build` command for example). Make the
script work like it’s a regular `nixos-rebuild` (just a couple of
options pre-bound).

The script was not very useful when it comes to custom arguments (as
they were not passed to the “default profile restoring” command). And I
was having hard time to remember the intended behavior. It’s easier to
just run a couple of commands manually instead.