Fix relative imports in nix-shell scripts (NixOS 24.11)
It’s actually a breaking change that came with 24.11 upgrade.
But this new change is actually great, since it imports stuff relative
to where the script is, not relative to current working directly.
Which means `channels/manage.raku` and `tell-a-secret.raku` can be ran
from anywhere now, just as before they were using `nixos-pin.nix`.