Tmuxed Alacritty: Improve robustness
Add retry mechanics for the generated name to handle race conditions.
If you press the terminal spawning hot key few times quick some of the
terminals will die because of the name conflict. First the script tries
to read the highest existic numeric session name, then increments, then
creates new session. 2 runs of the script can do read the existing names
and parallel, and end up with the same incremented number, trying to
create new session with the same name.

In order to fix this problem the script splits session creation into
first creating a detached session and then attaching. And it makes an
attempt to create the session, detect potential name conflict error,
then it retries to generate the name and do it again. Only after
successful session creation it attaches to it.

For the custom session name (not generated one) if the session with such
name already exists the script just attaches to it.